TS-567 Add support for jemalloc as well
git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/trunk@1178527 13f79535-47bb-0310-9956-ffa450edef68

+++ example/append-transform/append-transform.c
@@ -88,8 +88,8 @@ handle_transform(TSCont contp)
   TSVConn output_conn;
   TSVIO write_vio;
   MyData *data;
-  int towrite;
-  int avail;
+  int64_t towrite;
+  int64_t avail;
 
   /* Get the output connection where we'll write data to. */
   output_conn = TSTransformOutputVConnGet(contp);
